refactor: TypeScript configs and update build setup
- Split tsconfig into base, CJS, ESM, and test configs - Update
package.json exports and build scripts - Rename src/main.ts to
src/index.ts - Improve Jest config for ts-jest preset
