+++ tsconfig.tests.json
@@ -1,0 +1,9 @@
+{
+  "include": ["tests"],
+  "compilerOptions": {
+    "forceConsistentCasingInFileNames": true,
+    "strict": true,
+    "esModuleInterop": true,
+    "resolveJsonModule": true
+  }
+}
